update: move plugins/rm to aliases/avoid-mistakes

+++ aliases/avoid-mistakes.bash
@@ -1,6 +1,12 @@
 cite about-plugin
 about-plugin 'Some more alias to avoid making mistakes'
 
-alias rm='rm -i'
+if dotfiles_l.has command trash; then
+  # Safe "rm" command. Use trash if it exists. https://github.com/sindresorhus/trash
+  alias rm='trash'
+else
+  alias rm='rm -i'
+fi
+
 alias cp='cp -i'
 alias mv='mv -i'
